requirements.txt
google.generativeai 
gtts 
playsound
speechrecognition 
keyboard 
python-dotenv==0.19.0